==> rep_nosmooth/1-setup-model-table.R <==
#
#   Encode the row/column info for all models reflected in Tables 1
#   and 2.
#

library(tibble)
library(readr)
library(tidyr)
library(stringr)
library(dplyr)
library(rsample)
library(purrr)
library(yaml)
library(jsonlite)



# Table 1 cell definitions ------------------------------------------------
#
#   This defines the cells in Table 1
#

table1_row <- c("Base specification", "Terminal nodes", "Sample size",
                "Trees per forest", "Training/test sets 1",
                "Training/test sets 2", "Training/test sets 3",
                "Coding of DV 1", "Coding of DV 2")
table1_row <- factor(table1_row, levels = table1_row)

horizon <- c("1 month", "6 months")
horizon <- factor(horizon, levels = horizon)

# Aka the model specification (and average)
table1_column <- c("Escalation", "Quad", "Goldstein", "CAMEO", "Average")
table1_column <- factor(table1_column, levels = table1_column)

# This defines all cells in Table 1
table1 <- crossing(table = "Table 1", horizon, row = table1_row,
                   column = table1_column)


# Table 2 cell definitions ------------------------------------------------
#
#   This defines the cells in Table 2
#

horizon <- c("1 month", "6 months")
horizon <- factor(horizon, levels = horizon)

table2_column <- c("Escalation Only", "With PITF Predictors", "Weighted by PITF",
                   "PITF Split Population", "PITF Only")
table2_column <- factor(table2_column, levels = table2_column)

# This defines all cells in Table 2
table2 <- crossing(table = "Table 2", horizon, row = "Base specification",
                   column = table2_column)


# Combine table info ------------------------------------------------------

model_table <- bind_rows(table1, table2) %>%
  mutate(row = factor(row, levels = unique(row)),
         column = factor(column, levels = unique(column)))

# add a unique ID to the model table so that saving models is easier
model_table <- model_table %>%
  ungroup() %>%
  mutate(cell_id = 1:n())

# mark models that are not implemented as random forests
model_table <- model_table %>%
  mutate(non_RF = column %in% c("Average", "Weighted by PITF",
                                "PITF Split Population"))

model_table <- model_table %>%
  select(cell_id, everything())

# Model definitions / feature sets ----------------------------------------

escalation <- c(
  "gov_opp_low_level", "gov_reb_low_level", "opp_gov_low_level",
  "reb_gov_low_level", "gov_opp_nonviol_repression",
  "gov_reb_nonviol_repression", "gov_opp_accommodations",
  "gov_reb_accommodations", "reb_gov_demands", "opp_gov_demands"
)

quad <- c(
  "gov_opp_vercf","gov_reb_vercf","gov_opp_matcf","gov_reb_matcf",
  "opp_gov_vercf","reb_gov_vercf","opp_gov_matcf","reb_gov_matcf",
  "gov_opp_vercp","gov_reb_vercp","gov_opp_matcp","gov_reb_matcp",
  "opp_gov_vercp","reb_gov_vercp","opp_gov_matcp","reb_gov_matcp"
)

goldstein <- c(
  "gov_opp_gold","gov_reb_gold","opp_gov_gold","reb_gov_gold"
)

all_CAMEO <- c(
  "gov_opp_cameo_010","gov_opp_cameo_011","gov_opp_cameo_012",
  "gov_opp_cameo_013","gov_opp_cameo_014","gov_opp_cameo_015",
  "gov_opp_cameo_016","gov_opp_cameo_017","gov_opp_cameo_018",
  "gov_opp_cameo_019","gov_opp_cameo_020","gov_opp_cameo_021",
  "gov_opp_cameo_0211","gov_opp_cameo_0212","gov_opp_cameo_0213",
  "gov_opp_cameo_0214","gov_opp_cameo_022","gov_opp_cameo_023",
  "gov_opp_cameo_0231","gov_opp_cameo_0232","gov_opp_cameo_0233",
  "gov_opp_cameo_0234","gov_opp_cameo_024","gov_opp_cameo_0241",
  "gov_opp_cameo_0242","gov_opp_cameo_0243","gov_opp_cameo_0244",
  "gov_opp_cameo_025","gov_opp_cameo_0251","gov_opp_cameo_0252",
  "gov_opp_cameo_0253","gov_opp_cameo_0254","gov_opp_cameo_0255",
  "gov_opp_cameo_0256","gov_opp_cameo_026","gov_opp_cameo_027",
  "gov_opp_cameo_028","gov_opp_cameo_030","gov_opp_cameo_031",
  "gov_opp_cameo_0311","gov_opp_cameo_0312","gov_opp_cameo_0313",
  "gov_opp_cameo_0314","gov_opp_cameo_032","gov_opp_cameo_033",
  "gov_opp_cameo_0331","gov_opp_cameo_0332","gov_opp_cameo_0333",
  "gov_opp_cameo_0334","gov_opp_cameo_034","gov_opp_cameo_0341",
  "gov_opp_cameo_0342","gov_opp_cameo_0343", "gov_opp_cameo_0344",
  "gov_opp_cameo_035","gov_opp_cameo_0351", "gov_opp_cameo_0352",
  "gov_opp_cameo_0353", "gov_opp_cameo_0354", "gov_opp_cameo_0355",
  "gov_opp_cameo_0356", "gov_opp_cameo_036", "gov_opp_cameo_037",
  "gov_opp_cameo_038", "gov_opp_cameo_039", "gov_opp_cameo_040",
  "gov_opp_cameo_041", "gov_opp_cameo_042", "gov_opp_cameo_043",
  "gov_opp_cameo_044", "gov_opp_cameo_045", "gov_opp_cameo_046",
  "gov_opp_cameo_050", "gov_opp_cameo_051", "gov_opp_cameo_052",
  "gov_opp_cameo_053", "gov_opp_cameo_054", "gov_opp_cameo_055",
  "gov_opp_cameo_056", "gov_opp_cameo_057", "gov_opp_cameo_060",
  "gov_opp_cameo_061", "gov_opp_cameo_062", "gov_opp_cameo_063",
  "gov_opp_cameo_064", "gov_opp_cameo_070", "gov_opp_cameo_071",
  "gov_opp_cameo_072", "gov_opp_cameo_073", "gov_opp_cameo_074",
  "gov_opp_cameo_075", "gov_opp_cameo_080", "gov_opp_cameo_081",
  "gov_opp_cameo_0811", "gov_opp_cameo_0812", "gov_opp_cameo_0813",
  "gov_opp_cameo_0814", "gov_opp_cameo_082", "gov_opp_cameo_083",
  "gov_opp_cameo_0831", "gov_opp_cameo_0832", "gov_opp_cameo_0833",
  "gov_opp_cameo_0834", "gov_opp_cameo_084", "gov_opp_cameo_0841",
  "gov_opp_cameo_0842", "gov_opp_cameo_085", "gov_opp_cameo_086",
  "gov_opp_cameo_0861", "gov_opp_cameo_0862", "gov_opp_cameo_0863",
  "gov_opp_cameo_087","gov_opp_cameo_0871", "gov_opp_cameo_0872",
  "gov_opp_cameo_0873", "gov_opp_cameo_0874", "gov_opp_cameo_090",
  "gov_opp_cameo_091", "gov_opp_cameo_092", "gov_opp_cameo_093",
  "gov_opp_cameo_094", "gov_opp_cameo_100", "gov_opp_cameo_101",
  "gov_opp_cameo_1011", "gov_opp_cameo_1012","gov_opp_cameo_1013",
  "gov_opp_cameo_1014", "gov_opp_cameo_102", "gov_opp_cameo_103",
  "gov_opp_cameo_1031", "gov_opp_cameo_1032", "gov_opp_cameo_1033",
  "gov_opp_cameo_1034", "gov_opp_cameo_104", "gov_opp_cameo_1041",
  "gov_opp_cameo_1042", "gov_opp_cameo_1043", "gov_opp_cameo_1044",
  "gov_opp_cameo_105", "gov_opp_cameo_1051", "gov_opp_cameo_1052",
  "gov_opp_cameo_1053", "gov_opp_cameo_1054", "gov_opp_cameo_1055",
  "gov_opp_cameo_1056", "gov_opp_cameo_107", "gov_opp_cameo_108",
  "gov_opp_cameo_110", "gov_opp_cameo_111", "gov_opp_cameo_112",
  "gov_opp_cameo_1121", "gov_opp_cameo_1122", "gov_opp_cameo_1123",
  "gov_opp_cameo_1124", "gov_opp_cameo_1125", "gov_opp_cameo_113",
  "gov_opp_cameo_114", "gov_opp_cameo_115", "gov_opp_cameo_116",
  "gov_opp_cameo_120", "gov_opp_cameo_121", "gov_opp_cameo_1211",
  "gov_opp_cameo_1212", "gov_opp_cameo_122", "gov_opp_cameo_1221",
  "gov_opp_cameo_1222", "gov_opp_cameo_1223", "gov_opp_cameo_1224",
  "gov_opp_cameo_123", "gov_opp_cameo_1231", "gov_opp_cameo_1232",
  "gov_opp_cameo_1233", "gov_opp_cameo_1234", "gov_opp_cameo_124",
  "gov_opp_cameo_1241", "gov_opp_cameo_1242", "gov_opp_cameo_1243",
  "gov_opp_cameo_1244", "gov_opp_cameo_1245", "gov_opp_cameo_1246",
  "gov_opp_cameo_125", "gov_opp_cameo_126", "gov_opp_cameo_127",
  "gov_opp_cameo_128", "gov_opp_cameo_129", "gov_opp_cameo_130",
  "gov_opp_cameo_131", "gov_opp_cameo_1311", "gov_opp_cameo_1312",
  "gov_opp_cameo_1313", "gov_opp_cameo_132", "gov_opp_cameo_1321",
  "gov_opp_cameo_1322", "gov_opp_cameo_1323", "gov_opp_cameo_1324",
  "gov_opp_cameo_133", "gov_opp_cameo_134", "gov_opp_cameo_135",
  "gov_opp_cameo_136", "gov_opp_cameo_137", "gov_opp_cameo_138",
  "gov_opp_cameo_1381", "gov_opp_cameo_1382", "gov_opp_cameo_1383",
  "gov_opp_cameo_1384", "gov_opp_cameo_1385", "gov_opp_cameo_139",
  "gov_opp_cameo_140", "gov_opp_cameo_141", "gov_opp_cameo_1411",
  "gov_opp_cameo_1412", "gov_opp_cameo_1413", "gov_opp_cameo_1414",
  "gov_opp_cameo_142", "gov_opp_cameo_1421", "gov_opp_cameo_1422",
  "gov_opp_cameo_1423", "gov_opp_cameo_1424", "gov_opp_cameo_143",
  "gov_opp_cameo_1431", "gov_opp_cameo_1432", "gov_opp_cameo_1433",
  "gov_opp_cameo_1434", "gov_opp_cameo_144", "gov_opp_cameo_1441",
  "gov_opp_cameo_1442", "gov_opp_cameo_1443", "gov_opp_cameo_1444",
  "gov_opp_cameo_145", "gov_opp_cameo_1451", "gov_opp_cameo_1452",
  "gov_opp_cameo_1453", "gov_opp_cameo_1454", "gov_opp_cameo_150",
  "gov_opp_cameo_151", "gov_opp_cameo_152", "gov_opp_cameo_153",
  "gov_opp_cameo_154", "gov_opp_cameo_160", "gov_opp_cameo_161",
  "gov_opp_cameo_162", "gov_opp_cameo_1621", "gov_opp_cameo_1622",
  "gov_opp_cameo_1623", "gov_opp_cameo_163", "gov_opp_cameo_164",
  "gov_opp_cameo_165", "gov_opp_cameo_166", "gov_opp_cameo_1661",
  "gov_opp_cameo_1662", "gov_opp_cameo_1663", "gov_opp_cameo_170",
  "gov_opp_cameo_171", "gov_opp_cameo_1711", "gov_opp_cameo_1712",
  "gov_opp_cameo_172", "gov_opp_cameo_1721", "gov_opp_cameo_1722",
  "gov_opp_cameo_1723", "gov_opp_cameo_1724", "gov_opp_cameo_173",
  "gov_opp_cameo_174", "gov_opp_cameo_175", "gov_opp_cameo_180",
  "gov_opp_cameo_181", "gov_opp_cameo_182", "gov_opp_cameo_1821",
  "gov_opp_cameo_1822", "gov_opp_cameo_1823", "gov_opp_cameo_183",
  "gov_opp_cameo_1831", "gov_opp_cameo_1832", "gov_opp_cameo_1833",
  "gov_opp_cameo_184", "gov_opp_cameo_185", "gov_opp_cameo_186",
  "gov_opp_cameo_190", "gov_opp_cameo_191", "gov_opp_cameo_192",
  "gov_opp_cameo_193", "gov_opp_cameo_194", "gov_opp_cameo_195",
  "gov_opp_cameo_196", "gov_opp_cameo_200", "gov_opp_cameo_201",
  "gov_opp_cameo_202", "gov_opp_cameo_203", "gov_opp_cameo_204",
  "gov_opp_cameo_2041", "gov_opp_cameo_2042", "opp_gov_cameo_010",
  "opp_gov_cameo_011", "opp_gov_cameo_012", "opp_gov_cameo_013",
  "opp_gov_cameo_014", "opp_gov_cameo_015", "opp_gov_cameo_016",
  "opp_gov_cameo_017", "opp_gov_cameo_018", "opp_gov_cameo_019",
  "opp_gov_cameo_020", "opp_gov_cameo_021", "opp_gov_cameo_0211",
  "opp_gov_cameo_0212", "opp_gov_cameo_0213", "opp_gov_cameo_0214",
  "opp_gov_cameo_022", "opp_gov_cameo_023", "opp_gov_cameo_0231",
  "opp_gov_cameo_0232", "opp_gov_cameo_0233", "opp_gov_cameo_0234",
  "opp_gov_cameo_024", "opp_gov_cameo_0241", "opp_gov_cameo_0242",
  "opp_gov_cameo_0243", "opp_gov_cameo_0244", "opp_gov_cameo_025",
  "opp_gov_cameo_0251", "opp_gov_cameo_0252", "opp_gov_cameo_0253",
  "opp_gov_cameo_0254", "opp_gov_cameo_0255", "opp_gov_cameo_0256",
  "opp_gov_cameo_026", "opp_gov_cameo_027", "opp_gov_cameo_028",
  "opp_gov_cameo_030", "opp_gov_cameo_031", "opp_gov_cameo_0311",
  "opp_gov_cameo_0312", "opp_gov_cameo_0313", "opp_gov_cameo_0314",
  "opp_gov_cameo_032", "opp_gov_cameo_033", "opp_gov_cameo_0331",
  "opp_gov_cameo_0332", "opp_gov_cameo_0333", "opp_gov_cameo_0334",
  "opp_gov_cameo_034", "opp_gov_cameo_0341", "opp_gov_cameo_0342",
  "opp_gov_cameo_0343", "opp_gov_cameo_0344", "opp_gov_cameo_035",
  "opp_gov_cameo_0351", "opp_gov_cameo_0352", "opp_gov_cameo_0353",
  "opp_gov_cameo_0354", "opp_gov_cameo_0355", "opp_gov_cameo_0356",
  "opp_gov_cameo_036", "opp_gov_cameo_037", "opp_gov_cameo_038",
  "opp_gov_cameo_039", "opp_gov_cameo_040", "opp_gov_cameo_041",
  "opp_gov_cameo_042", "opp_gov_cameo_043", "opp_gov_cameo_044",
  "opp_gov_cameo_045", "opp_gov_cameo_046", "opp_gov_cameo_050",
  "opp_gov_cameo_051", "opp_gov_cameo_052", "opp_gov_cameo_053",
  "opp_gov_cameo_054", "opp_gov_cameo_055", "opp_gov_cameo_056",
  "opp_gov_cameo_057", "opp_gov_cameo_060", "opp_gov_cameo_061",
  "opp_gov_cameo_062", "opp_gov_cameo_063", "opp_gov_cameo_064",
  "opp_gov_cameo_070", "opp_gov_cameo_071", "opp_gov_cameo_072",
  "opp_gov_cameo_073", "opp_gov_cameo_074", "opp_gov_cameo_075",
  "opp_gov_cameo_080", "opp_gov_cameo_081", "opp_gov_cameo_0811",
  "opp_gov_cameo_0812", "opp_gov_cameo_0813", "opp_gov_cameo_0814",
  "opp_gov_cameo_082", "opp_gov_cameo_083", "opp_gov_cameo_0831",
  "opp_gov_cameo_0832", "opp_gov_cameo_0833", "opp_gov_cameo_0834",
  "opp_gov_cameo_084", "opp_gov_cameo_0841", "opp_gov_cameo_0842",
  "opp_gov_cameo_085", "opp_gov_cameo_086", "opp_gov_cameo_0861",
  "opp_gov_cameo_0862", "opp_gov_cameo_0863", "opp_gov_cameo_087",
  "opp_gov_cameo_0871", "opp_gov_cameo_0872", "opp_gov_cameo_0873",
  "opp_gov_cameo_0874", "opp_gov_cameo_090", "opp_gov_cameo_091",
  "opp_gov_cameo_092", "opp_gov_cameo_093", "opp_gov_cameo_094",
  "opp_gov_cameo_100", "opp_gov_cameo_101", "opp_gov_cameo_1011",
  "opp_gov_cameo_1012", "opp_gov_cameo_1013", "opp_gov_cameo_1014",
  "opp_gov_cameo_102", "opp_gov_cameo_103", "opp_gov_cameo_1031",
  "opp_gov_cameo_1032", "opp_gov_cameo_1033", "opp_gov_cameo_1034",
  "opp_gov_cameo_104", "opp_gov_cameo_1041", "opp_gov_cameo_1042",
  "opp_gov_cameo_1043", "opp_gov_cameo_1044", "opp_gov_cameo_105",
  "opp_gov_cameo_1051", "opp_gov_cameo_1052", "opp_gov_cameo_1053",
  "opp_gov_cameo_1054", "opp_gov_cameo_1055", "opp_gov_cameo_1056",
  "opp_gov_cameo_107", "opp_gov_cameo_108", "opp_gov_cameo_110",
  "opp_gov_cameo_111", "opp_gov_cameo_112", "opp_gov_cameo_1121",
  "opp_gov_cameo_1122", "opp_gov_cameo_1123", "opp_gov_cameo_1124",
  "opp_gov_cameo_1125", "opp_gov_cameo_113", "opp_gov_cameo_114",
  "opp_gov_cameo_115", "opp_gov_cameo_116", "opp_gov_cameo_120",
  "opp_gov_cameo_121", "opp_gov_cameo_1211", "opp_gov_cameo_1212",
  "opp_gov_cameo_122", "opp_gov_cameo_1221", "opp_gov_cameo_1222",
  "opp_gov_cameo_1223", "opp_gov_cameo_1224", "opp_gov_cameo_123",
  "opp_gov_cameo_1231", "opp_gov_cameo_1232", "opp_gov_cameo_1233",
  "opp_gov_cameo_1234", "opp_gov_cameo_124", "opp_gov_cameo_1241",
  "opp_gov_cameo_1242", "opp_gov_cameo_1243", "opp_gov_cameo_1244",
  "opp_gov_cameo_1245", "opp_gov_cameo_1246", "opp_gov_cameo_125",
  "opp_gov_cameo_126", "opp_gov_cameo_127", "opp_gov_cameo_128",
  "opp_gov_cameo_129", "opp_gov_cameo_130", "opp_gov_cameo_131",
  "opp_gov_cameo_1311", "opp_gov_cameo_1312", "opp_gov_cameo_1313",
  "opp_gov_cameo_132", "opp_gov_cameo_1321", "opp_gov_cameo_1322",
  "opp_gov_cameo_1323", "opp_gov_cameo_1324", "opp_gov_cameo_133",
  "opp_gov_cameo_134", "opp_gov_cameo_135", "opp_gov_cameo_136",
  "opp_gov_cameo_137", "opp_gov_cameo_138", "opp_gov_cameo_1381",
  "opp_gov_cameo_1382", "opp_gov_cameo_1383", "opp_gov_cameo_1384",
  "opp_gov_cameo_1385", "opp_gov_cameo_139", "opp_gov_cameo_140",
  "opp_gov_cameo_141", "opp_gov_cameo_1411", "opp_gov_cameo_1412",
  "opp_gov_cameo_1413", "opp_gov_cameo_1414", "opp_gov_cameo_142",
  "opp_gov_cameo_1421", "opp_gov_cameo_1422", "opp_gov_cameo_1423",
  "opp_gov_cameo_1424", "opp_gov_cameo_143", "opp_gov_cameo_1431",
  "opp_gov_cameo_1432", "opp_gov_cameo_1433", "opp_gov_cameo_1434",
  "opp_gov_cameo_144", "opp_gov_cameo_1441", "opp_gov_cameo_1442",
  "opp_gov_cameo_1443", "opp_gov_cameo_1444", "opp_gov_cameo_145",
  "opp_gov_cameo_1451", "opp_gov_cameo_1452", "opp_gov_cameo_1453",
  "opp_gov_cameo_1454", "opp_gov_cameo_150", "opp_gov_cameo_151",
  "opp_gov_cameo_152", "opp_gov_cameo_153", "opp_gov_cameo_154",
  "opp_gov_cameo_160", "opp_gov_cameo_161", "opp_gov_cameo_162",
  "opp_gov_cameo_1621", "opp_gov_cameo_1622", "opp_gov_cameo_1623",
  "opp_gov_cameo_163", "opp_gov_cameo_164", "opp_gov_cameo_165",
  "opp_gov_cameo_166", "opp_gov_cameo_1661", "opp_gov_cameo_1662",
  "opp_gov_cameo_1663", "opp_gov_cameo_170", "opp_gov_cameo_171",
  "opp_gov_cameo_1711", "opp_gov_cameo_1712", "opp_gov_cameo_172",
  "opp_gov_cameo_1721", "opp_gov_cameo_1722", "opp_gov_cameo_1723",
  "opp_gov_cameo_1724", "opp_gov_cameo_173", "opp_gov_cameo_174",
  "opp_gov_cameo_175", "opp_gov_cameo_180", "opp_gov_cameo_181",
  "opp_gov_cameo_182", "opp_gov_cameo_1821", "opp_gov_cameo_1822",
  "opp_gov_cameo_1823", "opp_gov_cameo_183", "opp_gov_cameo_1831",
  "opp_gov_cameo_1832", "opp_gov_cameo_1833", "opp_gov_cameo_184",
  "opp_gov_cameo_185", "opp_gov_cameo_186", "opp_gov_cameo_190",
  "opp_gov_cameo_191", "opp_gov_cameo_192", "opp_gov_cameo_193",
  "opp_gov_cameo_194", "opp_gov_cameo_195", "opp_gov_cameo_196",
  "opp_gov_cameo_200", "opp_gov_cameo_201", "opp_gov_cameo_202",
  "opp_gov_cameo_203", "opp_gov_cameo_204", "opp_gov_cameo_2041",
  "opp_gov_cameo_2042", "gov_reb_cameo_010", "gov_reb_cameo_011",
  "gov_reb_cameo_012", "gov_reb_cameo_013", "gov_reb_cameo_014",
  "gov_reb_cameo_015", "gov_reb_cameo_016", "gov_reb_cameo_017",
  "gov_reb_cameo_018", "gov_reb_cameo_019", "gov_reb_cameo_020",
  "gov_reb_cameo_021", "gov_reb_cameo_0211", "gov_reb_cameo_0212",
  "gov_reb_cameo_0213", "gov_reb_cameo_0214", "gov_reb_cameo_022",
  "gov_reb_cameo_023", "gov_reb_cameo_0231", "gov_reb_cameo_0232",
  "gov_reb_cameo_0233", "gov_reb_cameo_0234", "gov_reb_cameo_024",
  "gov_reb_cameo_0241", "gov_reb_cameo_0242", "gov_reb_cameo_0243",
  "gov_reb_cameo_0244", "gov_reb_cameo_025", "gov_reb_cameo_0251",
  "gov_reb_cameo_0252", "gov_reb_cameo_0253", "gov_reb_cameo_0254",
  "gov_reb_cameo_0255", "gov_reb_cameo_0256", "gov_reb_cameo_026",
  "gov_reb_cameo_027", "gov_reb_cameo_028", "gov_reb_cameo_030",
  "gov_reb_cameo_031", "gov_reb_cameo_0311", "gov_reb_cameo_0312",
  "gov_reb_cameo_0313", "gov_reb_cameo_0314", "gov_reb_cameo_032",
  "gov_reb_cameo_033", "gov_reb_cameo_0331", "gov_reb_cameo_0332",
  "gov_reb_cameo_0333", "gov_reb_cameo_0334", "gov_reb_cameo_034",
  "gov_reb_cameo_0341", "gov_reb_cameo_0342", "gov_reb_cameo_0343",
  "gov_reb_cameo_0344", "gov_reb_cameo_035", "gov_reb_cameo_0351",
  "gov_reb_cameo_0352", "gov_reb_cameo_0353", "gov_reb_cameo_0354",
  "gov_reb_cameo_0355", "gov_reb_cameo_0356", "gov_reb_cameo_036",
  "gov_reb_cameo_037", "gov_reb_cameo_038", "gov_reb_cameo_039",
  "gov_reb_cameo_040", "gov_reb_cameo_041", "gov_reb_cameo_042",
  "gov_reb_cameo_043", "gov_reb_cameo_044", "gov_reb_cameo_045",
  "gov_reb_cameo_046", "gov_reb_cameo_050", "gov_reb_cameo_051",
  "gov_reb_cameo_052", "gov_reb_cameo_053", "gov_reb_cameo_054",
  "gov_reb_cameo_055", "gov_reb_cameo_056", "gov_reb_cameo_057",
  "gov_reb_cameo_060", "gov_reb_cameo_061", "gov_reb_cameo_062",
  "gov_reb_cameo_063", "gov_reb_cameo_064", "gov_reb_cameo_070",
  "gov_reb_cameo_071", "gov_reb_cameo_072", "gov_reb_cameo_073",
  "gov_reb_cameo_074", "gov_reb_cameo_075", "gov_reb_cameo_080",
  "gov_reb_cameo_081", "gov_reb_cameo_0811", "gov_reb_cameo_0812",
  "gov_reb_cameo_0813", "gov_reb_cameo_0814", "gov_reb_cameo_082",
  "gov_reb_cameo_083", "gov_reb_cameo_0831", "gov_reb_cameo_0832",
  "gov_reb_cameo_0833", "gov_reb_cameo_0834", "gov_reb_cameo_084",
  "gov_reb_cameo_0841", "gov_reb_cameo_0842", "gov_reb_cameo_085",
  "gov_reb_cameo_086", "gov_reb_cameo_0861", "gov_reb_cameo_0862",
  "gov_reb_cameo_0863", "gov_reb_cameo_087", "gov_reb_cameo_0871",
  "gov_reb_cameo_0872", "gov_reb_cameo_0873", "gov_reb_cameo_0874",
  "gov_reb_cameo_090", "gov_reb_cameo_091", "gov_reb_cameo_092",
  "gov_reb_cameo_093", "gov_reb_cameo_094", "gov_reb_cameo_100",
  "gov_reb_cameo_101", "gov_reb_cameo_1011", "gov_reb_cameo_1012",
  "gov_reb_cameo_1013", "gov_reb_cameo_1014", "gov_reb_cameo_102",
  "gov_reb_cameo_103", "gov_reb_cameo_1031", "gov_reb_cameo_1032",
  "gov_reb_cameo_1033", "gov_reb_cameo_1034", "gov_reb_cameo_104",
  "gov_reb_cameo_1041", "gov_reb_cameo_1042", "gov_reb_cameo_1043",
  "gov_reb_cameo_1044", "gov_reb_cameo_105", "gov_reb_cameo_1051",
  "gov_reb_cameo_1052", "gov_reb_cameo_1053", "gov_reb_cameo_1054",
  "gov_reb_cameo_1055", "gov_reb_cameo_1056", "gov_reb_cameo_107",
  "gov_reb_cameo_108", "gov_reb_cameo_110", "gov_reb_cameo_111",
  "gov_reb_cameo_112", "gov_reb_cameo_1121", "gov_reb_cameo_1122",
  "gov_reb_cameo_1123", "gov_reb_cameo_1124", "gov_reb_cameo_1125",
  "gov_reb_cameo_113", "gov_reb_cameo_114", "gov_reb_cameo_115",
  "gov_reb_cameo_116", "gov_reb_cameo_120", "gov_reb_cameo_121",
  "gov_reb_cameo_1211", "gov_reb_cameo_1212", "gov_reb_cameo_122",
  "gov_reb_cameo_1221", "gov_reb_cameo_1222", "gov_reb_cameo_1223",
  "gov_reb_cameo_1224", "gov_reb_cameo_123", "gov_reb_cameo_1231",
  "gov_reb_cameo_1232", "gov_reb_cameo_1233", "gov_reb_cameo_1234",
  "gov_reb_cameo_124", "gov_reb_cameo_1241", "gov_reb_cameo_1242",
  "gov_reb_cameo_1243", "gov_reb_cameo_1244", "gov_reb_cameo_1245",
  "gov_reb_cameo_1246", "gov_reb_cameo_125", "gov_reb_cameo_126",
  "gov_reb_cameo_127", "gov_reb_cameo_128", "gov_reb_cameo_129",
  "gov_reb_cameo_130", "gov_reb_cameo_131", "gov_reb_cameo_1311",
  "gov_reb_cameo_1312", "gov_reb_cameo_1313", "gov_reb_cameo_132",
  "gov_reb_cameo_1321", "gov_reb_cameo_1322", "gov_reb_cameo_1323",
  "gov_reb_cameo_1324", "gov_reb_cameo_133", "gov_reb_cameo_134",
  "gov_reb_cameo_135", "gov_reb_cameo_136", "gov_reb_cameo_137",
  "gov_reb_cameo_138", "gov_reb_cameo_1381", "gov_reb_cameo_1382",
  "gov_reb_cameo_1383", "gov_reb_cameo_1384", "gov_reb_cameo_1385",
  "gov_reb_cameo_139", "gov_reb_cameo_140", "gov_reb_cameo_141",
  "gov_reb_cameo_1411", "gov_reb_cameo_1412", "gov_reb_cameo_1413",
  "gov_reb_cameo_1414", "gov_reb_cameo_142", "gov_reb_cameo_1421",
  "gov_reb_cameo_1422", "gov_reb_cameo_1423", "gov_reb_cameo_1424",
  "gov_reb_cameo_143", "gov_reb_cameo_1431", "gov_reb_cameo_1432",
  "gov_reb_cameo_1433", "gov_reb_cameo_1434", "gov_reb_cameo_144",
  "gov_reb_cameo_1441", "gov_reb_cameo_1442", "gov_reb_cameo_1443",
  "gov_reb_cameo_1444", "gov_reb_cameo_145", "gov_reb_cameo_1451",
  "gov_reb_cameo_1452", "gov_reb_cameo_1453", "gov_reb_cameo_1454",
  "gov_reb_cameo_150", "gov_reb_cameo_151", "gov_reb_cameo_152",
  "gov_reb_cameo_153", "gov_reb_cameo_154", "gov_reb_cameo_160",
  "gov_reb_cameo_161", "gov_reb_cameo_162", "gov_reb_cameo_1621",
  "gov_reb_cameo_1622", "gov_reb_cameo_1623", "gov_reb_cameo_163",
  "gov_reb_cameo_164", "gov_reb_cameo_165", "gov_reb_cameo_166",
  "gov_reb_cameo_1661", "gov_reb_cameo_1662", "gov_reb_cameo_1663",
  "gov_reb_cameo_170", "gov_reb_cameo_171", "gov_reb_cameo_1711",
  "gov_reb_cameo_1712", "gov_reb_cameo_172", "gov_reb_cameo_1721",
  "gov_reb_cameo_1722", "gov_reb_cameo_1723", "gov_reb_cameo_1724",
  "gov_reb_cameo_173", "gov_reb_cameo_174", "gov_reb_cameo_175",
  "gov_reb_cameo_180", "gov_reb_cameo_181", "gov_reb_cameo_182",
  "gov_reb_cameo_1821", "gov_reb_cameo_1822", "gov_reb_cameo_1823",
  "gov_reb_cameo_183", "gov_reb_cameo_1831", "gov_reb_cameo_1832",
  "gov_reb_cameo_1833", "gov_reb_cameo_184", "gov_reb_cameo_185",
  "gov_reb_cameo_186", "gov_reb_cameo_190", "gov_reb_cameo_191",
  "gov_reb_cameo_192", "gov_reb_cameo_193", "gov_reb_cameo_194",
  "gov_reb_cameo_195", "gov_reb_cameo_196", "gov_reb_cameo_200",
  "gov_reb_cameo_201", "gov_reb_cameo_202", "gov_reb_cameo_203",
  "gov_reb_cameo_204", "gov_reb_cameo_2041", "gov_reb_cameo_2042",
  "reb_gov_cameo_010", "reb_gov_cameo_011", "reb_gov_cameo_012",
  "reb_gov_cameo_013", "reb_gov_cameo_014", "reb_gov_cameo_015",
  "reb_gov_cameo_016", "reb_gov_cameo_017", "reb_gov_cameo_018",
  "reb_gov_cameo_019", "reb_gov_cameo_020", "reb_gov_cameo_021",
  "reb_gov_cameo_0211", "reb_gov_cameo_0212", "reb_gov_cameo_0213",
  "reb_gov_cameo_0214", "reb_gov_cameo_022", "reb_gov_cameo_023",
  "reb_gov_cameo_0231", "reb_gov_cameo_0232", "reb_gov_cameo_0233",
  "reb_gov_cameo_0234", "reb_gov_cameo_024", "reb_gov_cameo_0241",
  "reb_gov_cameo_0242", "reb_gov_cameo_0243", "reb_gov_cameo_0244",
  "reb_gov_cameo_025", "reb_gov_cameo_0251", "reb_gov_cameo_0252",
  "reb_gov_cameo_0253", "reb_gov_cameo_0254", "reb_gov_cameo_0255",
  "reb_gov_cameo_0256", "reb_gov_cameo_026", "reb_gov_cameo_027",
  "reb_gov_cameo_028", "reb_gov_cameo_030", "reb_gov_cameo_031",
  "reb_gov_cameo_0311", "reb_gov_cameo_0312", "reb_gov_cameo_0313",
  "reb_gov_cameo_0314", "reb_gov_cameo_032", "reb_gov_cameo_033",
  "reb_gov_cameo_0331", "reb_gov_cameo_0332", "reb_gov_cameo_0333",
  "reb_gov_cameo_0334", "reb_gov_cameo_034", "reb_gov_cameo_0341",
  "reb_gov_cameo_0342", "reb_gov_cameo_0343", "reb_gov_cameo_0344",
  "reb_gov_cameo_035", "reb_gov_cameo_0351", "reb_gov_cameo_0352",
  "reb_gov_cameo_0353", "reb_gov_cameo_0354", "reb_gov_cameo_0355",
  "reb_gov_cameo_0356", "reb_gov_cameo_036", "reb_gov_cameo_037",
  "reb_gov_cameo_038", "reb_gov_cameo_039", "reb_gov_cameo_040",
  "reb_gov_cameo_041", "reb_gov_cameo_042", "reb_gov_cameo_043",
  "reb_gov_cameo_044", "reb_gov_cameo_045", "reb_gov_cameo_046",
  "reb_gov_cameo_050", "reb_gov_cameo_051", "reb_gov_cameo_052",
  "reb_gov_cameo_053", "reb_gov_cameo_054", "reb_gov_cameo_055",
  "reb_gov_cameo_056", "reb_gov_cameo_057", "reb_gov_cameo_060",
  "reb_gov_cameo_061", "reb_gov_cameo_062", "reb_gov_cameo_063",
  "reb_gov_cameo_064", "reb_gov_cameo_070", "reb_gov_cameo_071",
  "reb_gov_cameo_072", "reb_gov_cameo_073", "reb_gov_cameo_074",
  "reb_gov_cameo_075", "reb_gov_cameo_080", "reb_gov_cameo_081",
  "reb_gov_cameo_0811", "reb_gov_cameo_0812", "reb_gov_cameo_0813",
  "reb_gov_cameo_0814", "reb_gov_cameo_082", "reb_gov_cameo_083",
  "reb_gov_cameo_0831", "reb_gov_cameo_0832", "reb_gov_cameo_0833",
  "reb_gov_cameo_0834", "reb_gov_cameo_084", "reb_gov_cameo_0841",
  "reb_gov_cameo_0842", "reb_gov_cameo_085", "reb_gov_cameo_086",
  "reb_gov_cameo_0861", "reb_gov_cameo_0862", "reb_gov_cameo_0863",
  "reb_gov_cameo_087", "reb_gov_cameo_0871", "reb_gov_cameo_0872",
  "reb_gov_cameo_0873", "reb_gov_cameo_0874", "reb_gov_cameo_090",
  "reb_gov_cameo_091", "reb_gov_cameo_092", "reb_gov_cameo_093",
  "reb_gov_cameo_094", "reb_gov_cameo_100", "reb_gov_cameo_101",
  "reb_gov_cameo_1011", "reb_gov_cameo_1012", "reb_gov_cameo_1013",
  "reb_gov_cameo_1014", "reb_gov_cameo_102", "reb_gov_cameo_103",
  "reb_gov_cameo_1031", "reb_gov_cameo_1032", "reb_gov_cameo_1033",
  "reb_gov_cameo_1034", "reb_gov_cameo_104", "reb_gov_cameo_1041",
  "reb_gov_cameo_1042", "reb_gov_cameo_1043", "reb_gov_cameo_1044",
  "reb_gov_cameo_105", "reb_gov_cameo_1051", "reb_gov_cameo_1052",
  "reb_gov_cameo_1053", "reb_gov_cameo_1054", "reb_gov_cameo_1055",
  "reb_gov_cameo_1056", "reb_gov_cameo_107", "reb_gov_cameo_108",
  "reb_gov_cameo_110", "reb_gov_cameo_111", "reb_gov_cameo_112",
  "reb_gov_cameo_1121", "reb_gov_cameo_1122", "reb_gov_cameo_1123",
  "reb_gov_cameo_1124", "reb_gov_cameo_1125", "reb_gov_cameo_113",
  "reb_gov_cameo_114", "reb_gov_cameo_115", "reb_gov_cameo_116",
  "reb_gov_cameo_120", "reb_gov_cameo_121", "reb_gov_cameo_1211",
  "reb_gov_cameo_1212", "reb_gov_cameo_122", "reb_gov_cameo_1221",
  "reb_gov_cameo_1222", "reb_gov_cameo_1223", "reb_gov_cameo_1224",
  "reb_gov_cameo_123", "reb_gov_cameo_1231", "reb_gov_cameo_1232",
  "reb_gov_cameo_1233", "reb_gov_cameo_1234", "reb_gov_cameo_124",
  "reb_gov_cameo_1241", "reb_gov_cameo_1242", "reb_gov_cameo_1243",
  "reb_gov_cameo_1244", "reb_gov_cameo_1245", "reb_gov_cameo_1246",
  "reb_gov_cameo_125", "reb_gov_cameo_126", "reb_gov_cameo_127",
  "reb_gov_cameo_128", "reb_gov_cameo_129", "reb_gov_cameo_130",
  "reb_gov_cameo_131", "reb_gov_cameo_1311", "reb_gov_cameo_1312",
  "reb_gov_cameo_1313", "reb_gov_cameo_132", "reb_gov_cameo_1321",
  "reb_gov_cameo_1322", "reb_gov_cameo_1323", "reb_gov_cameo_1324",
  "reb_gov_cameo_133", "reb_gov_cameo_134", "reb_gov_cameo_135",
  "reb_gov_cameo_136", "reb_gov_cameo_137", "reb_gov_cameo_138",
  "reb_gov_cameo_1381", "reb_gov_cameo_1382", "reb_gov_cameo_1383",
  "reb_gov_cameo_1384", "reb_gov_cameo_1385", "reb_gov_cameo_139",
  "reb_gov_cameo_140", "reb_gov_cameo_141", "reb_gov_cameo_1411",
  "reb_gov_cameo_1412", "reb_gov_cameo_1413", "reb_gov_cameo_1414",
  "reb_gov_cameo_142", "reb_gov_cameo_1421", "reb_gov_cameo_1422",
  "reb_gov_cameo_1423", "reb_gov_cameo_1424", "reb_gov_cameo_143",
  "reb_gov_cameo_1431", "reb_gov_cameo_1432", "reb_gov_cameo_1433",
  "reb_gov_cameo_1434", "reb_gov_cameo_144", "reb_gov_cameo_1441",
  "reb_gov_cameo_1442", "reb_gov_cameo_1443", "reb_gov_cameo_1444",
  "reb_gov_cameo_145", "reb_gov_cameo_1451", "reb_gov_cameo_1452",
  "reb_gov_cameo_1453", "reb_gov_cameo_1454", "reb_gov_cameo_150",
  "reb_gov_cameo_151", "reb_gov_cameo_152", "reb_gov_cameo_153",
  "reb_gov_cameo_154", "reb_gov_cameo_160", "reb_gov_cameo_161",
  "reb_gov_cameo_162", "reb_gov_cameo_1621", "reb_gov_cameo_1622",
  "reb_gov_cameo_1623", "reb_gov_cameo_163", "reb_gov_cameo_164",
  "reb_gov_cameo_165", "reb_gov_cameo_166", "reb_gov_cameo_1661",
  "reb_gov_cameo_1662", "reb_gov_cameo_1663", "reb_gov_cameo_170",
  "reb_gov_cameo_171", "reb_gov_cameo_1711", "reb_gov_cameo_1712",
  "reb_gov_cameo_172", "reb_gov_cameo_1721", "reb_gov_cameo_1722",
  "reb_gov_cameo_1723", "reb_gov_cameo_1724", "reb_gov_cameo_173",
  "reb_gov_cameo_174", "reb_gov_cameo_175", "reb_gov_cameo_180",
  "reb_gov_cameo_181", "reb_gov_cameo_182", "reb_gov_cameo_1821",
  "reb_gov_cameo_1822", "reb_gov_cameo_1823", "reb_gov_cameo_183",
  "reb_gov_cameo_1831", "reb_gov_cameo_1832", "reb_gov_cameo_1833",
  "reb_gov_cameo_184", "reb_gov_cameo_185", "reb_gov_cameo_186",
  "reb_gov_cameo_190", "reb_gov_cameo_191", "reb_gov_cameo_192",
  "reb_gov_cameo_193", "reb_gov_cameo_194", "reb_gov_cameo_195",
  "reb_gov_cameo_196", "reb_gov_cameo_200", "reb_gov_cameo_201",
  "reb_gov_cameo_202", "reb_gov_cameo_203", "reb_gov_cameo_204",
  "reb_gov_cameo_2041", "reb_gov_cameo_2042")

PITF <- c(
  "region_africa","region_easia","region_europe","region_latam","region_neast",
  "disp4cat_minus1","faction_minus1","bnkv100cat_minus1","lncimr_minus1",
  "lntpopx_minus1"
)

# the names of this list should match the Table 1 and 2 column names, which
# encode the specifications used (or model type, like Average)
specs <- list(
  Escalation              = escalation,
  Quad                    = quad,
  Goldstein               = goldstein,
  CAMEO                   = all_CAMEO,
  Average                 = NA,
  `Escalation Only`       = escalation,
  `With PITF Predictors`  = c(escalation, PITF),
  `Weighted by PITF`      = escalation,
  `PITF Split Population` = escalation,
  `PITF Only`             = PITF
)

stopifnot(
  all(table1_column %in% names(specs)),
  all(table2_column %in% names(specs))
)

feature_specs <- specs


# Encode table row / robustness check info --------------------------------
#
#   The rows in Table 1 and Table 2, aside from the forecast horizon / data
#   aggregation level, alter several dimensions:
#
#     - the DV used
#     - the random forest hyperparameter settings used
#     - the train/test split
#
#   Encode all of these choices in three lists that we can use to retrieve
#   the relevant settings for a row in table 1.
#

# Dependent variable
# use like: dv_specs[[table row]]

dv_specs <- list(
  `Base specification`   = "incidence_civil_ns_plus1",
  `Terminal nodes`       = "incidence_civil_ns_plus1",
  `Sample size`          = "incidence_civil_ns_plus1",
  `Trees per forest`     = "incidence_civil_ns_plus1",
  `Training/test sets 1` = "incidence_civil_ns_plus1",
  `Training/test sets 2` = "incidence_civil_ns_plus1",
  `Training/test sets 3` = "incidence_civil_ns_plus1",
  `Coding of DV 1`       = "incidence_civil_ns_alt1_plus1",
  `Coding of DV 2`       = "incidence_civil_ns_alt2_plus1"
)

# Hyperparameter settings
# use like: hp_settings[[table row]][[hyper parameter]]

basic_hp_settings <- list(
  ntree    = 100000L,
  maxnodes = 5L,
  sampsize = 100L,
  replace  = FALSE
)

hp_settings <- list(
  `Base specification`   = basic_hp_settings,
  `Terminal nodes`       = basic_hp_settings %>% list_modify(maxnodes = 10L),
  `Sample size`          = basic_hp_settings %>% list_modify(sampsize = 500L),
  `Trees per forest`     = basic_hp_settings %>% list_modify(ntree = 1e6L),
  `Training/test sets 1` = basic_hp_settings,
  `Training/test sets 2` = basic_hp_settings,
  `Training/test sets 3` = basic_hp_settings,
  `Coding of DV 1`       = basic_hp_settings,
  `Coding of DV 2`       = basic_hp_settings
)

# Train/test split
# This one depends on
## RM: This one depends on what???

train_end_year <- list(
  `Base specification`   = 2007,
  `Terminal nodes`       = 2007,
  `Sample size`          = 2007,
  `Trees per forest`     = 2007,
  `Training/test sets 1` = 2008,
  `Training/test sets 2` = 2009,
  `Training/test sets 3` = 2010,
  `Coding of DV 1`       = 2007,
  `Coding of DV 2`       = 2007
) %>%
  lapply(as.integer)

stopifnot(
  all(table1_row %in% names(dv_specs)),
  all(table1_row %in% names(hp_settings)),
  all(table1_row %in% names(train_end_year))
)



# Save model definitions --------------------------------------------------
#
#   To the extent possible, save these in a human readable format. This doesn't
#   work well with the model table and features specs because they have too
#   many elements, but the others are ok as YAML. The feature specs are ok as
#   JSON.
#

dir.create(here("rep_nosmooth/output/model-definitions"),
           recursive = TRUE, showWarnings = FALSE)

write_rds(model_table,
          here("rep_nosmooth/output/model-definitions/model-table.rds"))
write_json(feature_specs,
           here("rep_nosmooth/output/model-definitions/feature-specs.json"),
           pretty = TRUE, auto_unbox = TRUE)
write_yaml(hp_settings,
           here("rep_nosmooth/output/model-definitions/hp-settings.yaml"))
write_yaml(dv_specs,
           here("rep_nosmooth/output/model-definitions/dv-specs.yaml"))
write_yaml(train_end_year,
           here("rep_nosmooth/output/model-definitions/train-end-year.yaml"))
